+++ lib/bolt.rb
@@ -1,4 +1,3 @@
 module Bolt
   require 'bolt/executor'
-  require 'bolt/node'
 end
