+++ lib/bolt/cli.rb
@@ -9,7 +9,6 @@ require 'bolt/error'
 require 'bolt/executor'
 require 'bolt/inventory'
 require 'bolt/logger'
-require 'bolt/node'
 require 'bolt/outputter'
 require 'bolt/pal'
 require 'bolt/target'
