+++ lib/bolt/executor.rb
@@ -4,17 +4,25 @@ require 'logging'
 require 'bolt/result'
 require 'bolt/config'
 require 'bolt/notifier'
-require 'bolt/node'
 require 'bolt/result_set'
+require 'bolt/transport/ssh'
+require 'bolt/transport/winrm'
+require 'bolt/transport/orch'
 
 module Bolt
   class Executor
-    attr_reader :noop
+    attr_reader :noop, :transports
     attr_accessor :run_as
 
     def initialize(config = Bolt::Config.new, noop = nil, plan_logging = false)
       @config = config
       @logger = Logging.logger[self]
+
+      @transports = {
+        'ssh' => Bolt::Transport::SSH.new(config[:transports][:ssh] || {}),
+        'winrm' => Bolt::Transport::WinRM.new(config[:transports][:winrm] || {}),
+        'pcp' => Bolt::Transport::Orch.new(config[:transports][:pcp] || {})
+      }
 
       # If a specific elevated log level has been requested, honor that.
       # Otherwise, escalate the log level to "info" if running in plan mode, so
@@ -26,39 +34,22 @@ module Bolt
       @notifier = Bolt::Notifier.new
     end
 
-    def from_targets(targets)
-      targets.map do |target|
-        Bolt::Node.from_target(target)
-      end
-    end
-    private :from_targets
-
-    def on(nodes, callback = nil)
+    def on(targets, callback = nil)
       results = Concurrent::Array.new
 
-      poolsize = [nodes.length, @config[:concurrency]].min
+      poolsize = [targets.length, @config[:concurrency]].min
       pool = Concurrent::FixedThreadPool.new(poolsize)
       @logger.debug { "Started with #{poolsize} thread(s)" }
 
-      nodes.map(&:class).uniq.each do |klass|
-        klass.initialize_transport(@logger)
-      end
-
-      nodes.each { |node|
+      targets.each { |target|
         pool.post do
           result =
             begin
-              @notifier.notify(callback, type: :node_start, target: node.target) if callback
-              node.connect
-              yield node
+              transport = @transports[target.protocol || 'ssh']
+              @notifier.notify(callback, type: :node_start, target: target) if callback
+              yield transport, target
             rescue StandardError => ex
-              Bolt::Result.from_exception(node.target, ex)
-            ensure
-              begin
-                node.disconnect
-              rescue StandardError => ex
-                @logger.info("Failed to close connection to #{node.uri} : #{ex.message}")
-              end
+              Bolt::Result.from_exception(target, ex)
             end
           results.concat([result])
           @notifier.notify(callback, type: :node_result, result: result) if callback
@@ -81,8 +72,8 @@ module Bolt
     end
     private :summary
 
-    def get_run_as(node, options)
-      if node.run_as.nil? && run_as
+    def get_run_as(target, options)
+      if target.options[:run_as].nil? && run_as
         { '_run_as' => run_as }.merge(options)
       else
         options
@@ -90,78 +81,74 @@ module Bolt
     end
     private :get_run_as
 
-    def with_exception_handling(node)
+    def with_exception_handling(target)
       yield
     rescue StandardError => e
-      Bolt::Result.from_exception(node.target, e)
+      Bolt::Result.from_exception(target, e)
     end
     private :with_exception_handling
 
     def run_command(targets, command, options = {})
-      nodes = from_targets(targets)
-      @logger.info("Starting command run '#{command}' on #{nodes.map(&:uri)}")
+      @logger.info("Starting command run '#{command}' on #{targets.map(&:uri)}")
       callback = block_given? ? Proc.new : nil
 
-      r = on(nodes, callback) do |node|
-        @logger.debug("Running command '#{command}' on #{node.uri}")
-        node_result = with_exception_handling(node) do
-          node.run_command(command, get_run_as(node, options))
+      r = on(targets, callback) do |transport, target|
+        @logger.debug("Running command '#{command}' on #{target.uri}")
+        target_result = with_exception_handling(target) do
+          transport.run_command(target, command, get_run_as(target, options))
         end
-        @logger.debug("Result on #{node.uri}: #{JSON.dump(node_result.value)}")
-        node_result
+        @logger.debug("Result on #{target.uri}: #{JSON.dump(target_result.value)}")
+        target_result
       end
       @logger.info(summary('command', command, r))
       r
     end
 
     def run_script(targets, script, arguments, options = {})
-      nodes = from_targets(targets)
-      @logger.info("Starting script run #{script} on #{nodes.map(&:uri)}")
+      @logger.info("Starting script run #{script} on #{targets.map(&:uri)}")
       @logger.debug("Arguments: #{arguments}")
       callback = block_given? ? Proc.new : nil
 
-      r = on(nodes, callback) do |node|
-        @logger.debug { "Running script '#{script}' on #{node.uri}" }
-        node_result = with_exception_handling(node) do
-          node.run_script(script, arguments, get_run_as(node, options))
+      r = on(targets, callback) do |transport, target|
+        @logger.debug { "Running script '#{script}' on #{target.uri}" }
+        target_result = with_exception_handling(target) do
+          transport.run_script(target, script, arguments, get_run_as(target, options))
         end
-        @logger.debug("Result on #{node.uri}: #{JSON.dump(node_result.value)}")
-        node_result
+        @logger.debug("Result on #{target.uri}: #{JSON.dump(target_result.value)}")
+        target_result
       end
       @logger.info(summary('script', script, r))
       r
     end
 
     def run_task(targets, task, input_method, arguments, options = {})
-      nodes = from_targets(targets)
-      @logger.info("Starting task #{task} on #{nodes.map(&:uri)}")
+      @logger.info("Starting task #{task} on #{targets.map(&:uri)}")
       @logger.debug("Arguments: #{arguments} Input method: #{input_method}")
       callback = block_given? ? Proc.new : nil
 
-      r = on(nodes, callback) do |node|
-        @logger.debug { "Running task run '#{task}' on #{node.uri}" }
-        node_result = with_exception_handling(node) do
-          node.run_task(task, input_method, arguments, get_run_as(node, options))
+      r = on(targets, callback) do |transport, target|
+        @logger.debug { "Running task run '#{task}' on #{target.uri}" }
+        target_result = with_exception_handling(target) do
+          transport.run_task(target, task, input_method, arguments, get_run_as(target, options))
         end
-        @logger.debug("Result on #{node.uri}: #{JSON.dump(node_result.value)}")
-        node_result
+        @logger.debug("Result on #{target.uri}: #{JSON.dump(target_result.value)}")
+        target_result
       end
       @logger.info(summary('task', task, r))
       r
     end
 
     def file_upload(targets, source, destination, options = {})
-      nodes = from_targets(targets)
-      @logger.info("Starting file upload from #{source} to #{destination} on #{nodes.map(&:uri)}")
+      @logger.info("Starting file upload from #{source} to #{destination} on #{targets.map(&:uri)}")
       callback = block_given? ? Proc.new : nil
 
-      r = on(nodes, callback) do |node|
-        @logger.debug { "Uploading: '#{source}' to #{destination} on #{node.uri}" }
-        node_result = with_exception_handling(node) do
-          node.upload(source, destination, options)
+      r = on(targets, callback) do |transport, target|
+        @logger.debug { "Uploading: '#{source}' to #{destination} on #{target.uri}" }
+        target_result = with_exception_handling(target) do
+          transport.upload(target, source, destination, options)
         end
-        @logger.debug("Result on #{node.uri}: #{JSON.dump(node_result.value)}")
-        node_result
+        @logger.debug("Result on #{target.uri}: #{JSON.dump(target_result.value)}")
+        target_result
       end
       @logger.info(summary('upload', source, r))
       r
