+++ spec/bolt/executor_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 require 'bolt/executor'
 
 describe "Bolt::Executor" do
-  let(:config) { Bolt::Config.new }
+  let(:config) { Bolt::Config.new(concurrency: 1) }
   let(:executor) { Bolt::Executor.new(config) }
   let(:command) { "hostname" }
   let(:script) { '/path/to/script.sh' }
@@ -19,28 +19,21 @@ describe "Bolt::Executor" do
     { type: :node_result, result: result }
   end
 
-  def mock_node(name, target, run_as)
-    double(name, class: transport, connect: nil, disconnect: nil, uri: name, target: target, run_as: run_as)
-  end
-
-  def mock_node_results(run_as = nil)
+  def mock_node_results(_run_as = nil)
     {
-      mock_node('node1', targets[0], run_as) => Bolt::Result.new(targets[0]),
-      mock_node('node2', targets[1], run_as) => Bolt::Result.new(targets[0])
+      targets[0] => Bolt::Result.new(targets[0]),
+      targets[1] => Bolt::Result.new(targets[1])
     }
   end
 
-  let(:targets) { [double('target1'), double('target2')] }
+  let(:targets) { [Bolt::Target.new("target1"), Bolt::Target.new("target2")] }
   let(:node_results) { mock_node_results }
-
-  before(:each) do
-    allow(executor).to receive(:from_targets).with(targets).and_return(node_results.keys)
-  end
+  let(:ssh) { executor.transports['ssh'] }
 
   context 'running a command' do
     it 'executes on all nodes' do
-      node_results.each do |node, result|
-        expect(node).to receive(:run_command).with(command, {}).and_return(result)
+      node_results.each do |target, result|
+        expect(ssh).to receive(:run_command).with(target, command, {}).and_return(result)
       end
 
       executor.run_command(targets, command, {})
@@ -48,20 +41,23 @@ describe "Bolt::Executor" do
 
     it 'passes _run_as' do
       executor.run_as = 'foo'
-      node_results.each do |node, result|
-        expect(node).to receive(:run_command).with(command, '_run_as' => 'foo').and_return(result)
+      node_results.each do |target, result|
+        expect(ssh).to receive(:run_command).with(target, command, '_run_as' => 'foo').and_return(result)
       end
 
       executor.run_command(targets, command)
     end
 
     context 'nodes with run_as' do
-      let(:node_results) { mock_node_results('foo') }
+      let(:targets) {
+        [Bolt::Target.new("target1", run_as: 'foo'),
+         Bolt::Target.new("target2", run_as: 'foo')]
+      }
 
       it 'does not pass _run_as' do
         executor.run_as = 'foo'
-        node_results.each do |node, result|
-          expect(node).to receive(:run_command).with(command, {}).and_return(result)
+        node_results.each do |target, result|
+          expect(ssh).to receive(:run_command).with(target, command, {}).and_return(result)
         end
 
         executor.run_command(targets, command)
@@ -69,8 +65,8 @@ describe "Bolt::Executor" do
     end
 
     it "yields each result" do
-      node_results.each do |node, result|
-        expect(node).to receive(:run_command).with(command, {}).and_return(result)
+      node_results.each do |target, result|
+        expect(ssh).to receive(:run_command).with(target, command, {}).and_return(result)
       end
 
       results = []
@@ -78,15 +74,15 @@ describe "Bolt::Executor" do
         results << result
       end
 
-      node_results.each do |node, result|
+      node_results.each do |target, result|
         expect(results).to include(success_event(result))
-        expect(results).to include(start_event(node.target))
+        expect(results).to include(start_event(target))
       end
     end
 
     it 'catches errors' do
-      node_results.each_key do |node|
-        expect(node).to receive(:run_command).with(command, {}).and_raise(Bolt::Error, 'failed', 'my-exception')
+      node_results.each_key do |target|
+        expect(ssh).to receive(:run_command).with(target, command, {}).and_raise(Bolt::Error, 'failed', 'my-exception')
       end
 
       executor.run_command(targets, command) do |result|
@@ -98,8 +94,8 @@ describe "Bolt::Executor" do
 
   context 'executes running a script' do
     it "on all nodes" do
-      node_results.each do |node, result|
-        expect(node).to receive(:run_script).with(script, [], {}).and_return(result)
+      node_results.each do |target, result|
+        expect(ssh).to receive(:run_script).with(target, script, [], {}).and_return(result)
       end
 
       results = executor.run_script(targets, script, [], {})
@@ -110,8 +106,8 @@ describe "Bolt::Executor" do
 
     it 'passes _run_as' do
       executor.run_as = 'foo'
-      node_results.each do |node, result|
-        expect(node).to receive(:run_script).with(script, [], '_run_as' => 'foo').and_return(result)
+      node_results.each do |target, result|
+        expect(ssh).to receive(:run_script).with(target, script, [], '_run_as' => 'foo').and_return(result)
       end
 
       results = executor.run_script(targets, script, [])
@@ -121,12 +117,15 @@ describe "Bolt::Executor" do
     end
 
     context 'nodes with run_as' do
-      let(:node_results) { mock_node_results('foo') }
+      let(:targets) {
+        [Bolt::Target.new("target1", run_as: 'foo'),
+         Bolt::Target.new("target2", run_as: 'foo')]
+      }
 
       it 'does not pass _run_as' do
         executor.run_as = 'foo'
-        node_results.each do |node, result|
-          expect(node).to receive(:run_script).with(script, [], {}).and_return(result)
+        node_results.each do |target, result|
+          expect(ssh).to receive(:run_script).with(target, script, [], {}).and_return(result)
         end
 
         results = executor.run_script(targets, script, [])
@@ -137,8 +136,8 @@ describe "Bolt::Executor" do
     end
 
     it "yields each result" do
-      node_results.each do |node, result|
-        expect(node).to receive(:run_script).with(script, [], {}).and_return(result)
+      node_results.each do |target, result|
+        expect(ssh).to receive(:run_script).with(target, script, [], {}).and_return(result)
       end
 
       results = []
@@ -146,15 +145,18 @@ describe "Bolt::Executor" do
         results << result
       end
 
-      node_results.each do |node, result|
+      node_results.each do |target, result|
         expect(results).to include(success_event(result))
-        expect(results).to include(start_event(node.target))
+        expect(results).to include(start_event(target))
       end
     end
 
     it 'catches errors' do
-      node_results.each_key do |node|
-        expect(node).to receive(:run_script).with(script, [], {}).and_raise(Bolt::Error, 'failed', 'my-exception')
+      node_results.each_key do |target|
+        expect(ssh)
+          .to receive(:run_script)
+          .with(target, script, [], {})
+          .and_raise(Bolt::Error, 'failed', 'my-exception')
       end
 
       executor.run_script(targets, script, []) do |result|
@@ -166,10 +168,10 @@ describe "Bolt::Executor" do
 
   context 'running a task' do
     it "executes on all nodes" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:run_task)
-          .with(task, 'both', task_arguments, {})
+          .with(target, task, 'both', task_arguments, {})
           .and_return(result)
       end
 
@@ -181,10 +183,10 @@ describe "Bolt::Executor" do
 
     it 'passes _run_as' do
       executor.run_as = 'foo'
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:run_task)
-          .with(task, 'both', task_arguments, '_run_as' => 'foo')
+          .with(target, task, 'both', task_arguments, '_run_as' => 'foo')
           .and_return(result)
       end
 
@@ -195,14 +197,17 @@ describe "Bolt::Executor" do
     end
 
     context 'nodes with run_as' do
-      let(:node_results) { mock_node_results('foo') }
+      let(:targets) {
+        [Bolt::Target.new("target1", run_as: 'foo'),
+         Bolt::Target.new("target2", run_as: 'foo')]
+      }
 
       it 'does not pass _run_as' do
         executor.run_as = 'foo'
-        node_results.each do |node, result|
-          expect(node)
+        node_results.each do |target, result|
+          expect(ssh)
             .to receive(:run_task)
-            .with(task, 'both', task_arguments, {})
+            .with(target, task, 'both', task_arguments, {})
             .and_return(result)
         end
 
@@ -214,10 +219,10 @@ describe "Bolt::Executor" do
     end
 
     it "yields each result" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:run_task)
-          .with(task, 'both', task_arguments, {})
+          .with(target, task, 'both', task_arguments, {})
           .and_return(result)
       end
 
@@ -225,17 +230,17 @@ describe "Bolt::Executor" do
       executor.run_task(targets, task, 'both', task_arguments) do |result|
         results << result
       end
-      node_results.each do |node, result|
+      node_results.each do |target, result|
         expect(results).to include(success_event(result))
-        expect(results).to include(start_event(node.target))
+        expect(results).to include(start_event(target))
       end
     end
 
     it 'catches errors' do
-      node_results.each_key do |node|
-        expect(node)
+      node_results.each_key do |target|
+        expect(ssh)
           .to receive(:run_task)
-          .with(task, 'both', task_arguments, {})
+          .with(target, task, 'both', task_arguments, {})
           .and_raise(Bolt::Error, 'failed', 'my-exception')
       end
 
@@ -248,10 +253,10 @@ describe "Bolt::Executor" do
 
   context 'uploading a file' do
     it "executes on all nodes" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:upload)
-          .with(script, dest, {})
+          .with(target, script, dest, {})
           .and_return(result)
       end
 
@@ -262,10 +267,10 @@ describe "Bolt::Executor" do
     end
 
     it "yields each result" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:upload)
-          .with(script, dest, {})
+          .with(target, script, dest, {})
           .and_return(result)
       end
 
@@ -273,17 +278,17 @@ describe "Bolt::Executor" do
       executor.file_upload(targets, script, dest) do |result|
         results << result
       end
-      node_results.each do |node, result|
+      node_results.each do |target, result|
         expect(results).to include(success_event(result))
-        expect(results).to include(start_event(node.target))
+        expect(results).to include(start_event(target))
       end
     end
 
     it 'catches errors' do
-      node_results.each_key do |node|
-        expect(node)
+      node_results.each_key do |target|
+        expect(ssh)
           .to receive(:upload)
-          .with(script, dest, {})
+          .with(target, script, dest, {})
           .and_raise(Bolt::Error, 'failed', 'my-exception')
       end
 
@@ -295,9 +300,9 @@ describe "Bolt::Executor" do
   end
 
   it "returns an error result" do
-    node_results.each_key do |node|
-      expect(node)
-        .to receive(:connect)
+    node_results.each_key do |_target|
+      expect(ssh)
+        .to receive(:with_connection)
         .and_raise(
           Bolt::Node::ConnectError.new('Authentication failed', 'AUTH_ERROR')
         )
@@ -310,10 +315,8 @@ describe "Bolt::Executor" do
   end
 
   it "returns an exception result if the connect raises an unhandled error" do
-    logger = double('logger', error: nil)
-    node_results.each_key do |node|
-      allow(node).to receive(:logger).and_return(logger)
-      expect(node).to receive(:connect).and_raise("reset")
+    node_results.each_key do |_target|
+      expect(ssh).to receive(:with_connection).and_raise("reset")
     end
 
     results = executor.run_command(targets, command)
@@ -324,13 +327,13 @@ describe "Bolt::Executor" do
 
   context "When running a plan" do
     let(:executor) { Bolt::Executor.new(config, nil, true) }
-    let(:nodes_string) { node_results.map(&:first).map(&:uri) }
+    let(:nodes_string) { results.map(&:first).map(&:uri) }
 
     it "logs commands" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:run_command)
-          .with(command, {})
+          .with(target, command, {})
           .and_return(result)
       end
 
@@ -341,10 +344,10 @@ describe "Bolt::Executor" do
     end
 
     it "logs scripts" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:run_script)
-          .with(script, [], {})
+          .with(target, script, [], {})
           .and_return(result)
       end
 
@@ -355,10 +358,10 @@ describe "Bolt::Executor" do
     end
 
     it "logs tasks" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:run_task)
-          .with(task, 'both', task_arguments, {})
+          .with(target, task, 'both', task_arguments, {})
           .and_return(result)
       end
 
@@ -369,10 +372,10 @@ describe "Bolt::Executor" do
     end
 
     it "logs uploads" do
-      node_results.each do |node, result|
-        expect(node)
+      node_results.each do |target, result|
+        expect(ssh)
           .to receive(:upload)
-          .with(script, dest, {})
+          .with(target, script, dest, {})
           .and_return(result)
       end
