+++ spec/bolt/transport/ssh_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
+require 'net/ssh'
 require 'bolt_spec/errors'
 require 'bolt_spec/files'
-require 'bolt/node'
-require 'bolt/node/ssh'
+require 'bolt/transport/ssh'
 require 'bolt/config'
 
-describe Bolt::SSH do
+describe Bolt::Transport::SSH do
   include BoltSpec::Errors
   include BoltSpec::Files
 
@@ -21,7 +21,7 @@ describe Bolt::SSH do
   let(:command) { "pwd" }
   let(:config) { mk_config(user: user, password: password) }
   let(:no_host_key_check) { mk_config(host_key_check: false, user: user, password: password) }
-  let(:ssh) { Bolt::SSH.new(target) }
+  let(:ssh) { Bolt::Transport::SSH.new(config) }
   let(:echo_script) { <<BASH }
 for var in "$@"
 do
@@ -29,9 +29,11 @@ do
 done
 BASH
 
-  def target(h: hostname, p: port, conf: config)
+  def make_target(h: hostname, p: port, conf: config)
     Bolt::Target.new("#{h}:#{p}").update_conf(conf.transport_conf)
   end
+
+  let(:target) { make_target }
 
   def result_value(stdout = nil, stderr = nil, exit_code = 0)
     { 'stdout' => stdout || '',
@@ -48,12 +50,10 @@ BASH
               hash_including(
                 verify_host_key: instance_of(Net::SSH::Verifiers::Secure)
               ))
-      ssh.connect
+      ssh.with_connection(target) {}
     end
 
     it "downgrades to lenient if host_key_check is false" do
-      ssh = Bolt::SSH.new(target(conf: no_host_key_check))
-
       allow(Net::SSH)
         .to receive(:start)
         .with(anything,
@@ -61,20 +61,18 @@ BASH
               hash_including(
                 verify_host_key: instance_of(Net::SSH::Verifiers::Lenient)
               ))
-      ssh.connect
+      ssh.with_connection(make_target(conf: no_host_key_check)) {}
     end
 
     it "rejects the connection if host key verification fails" do
       expect_node_error(Bolt::Node::ConnectError,
                         'HOST_KEY_ERROR',
                         /Host key verification failed/) do
-        ssh.connect
+        ssh.with_connection(target) {}
       end
     end
 
     it "raises ConnectError if authentication fails" do
-      ssh = Bolt::SSH.new(target(conf: no_host_key_check))
-
       allow(Net::SSH)
         .to receive(:start)
         .and_raise(Net::SSH::AuthenticationFailed,
@@ -82,18 +80,17 @@ BASH
       expect_node_error(Bolt::Node::ConnectError,
                         'AUTH_ERROR',
                         /Authentication failed for foo@bar.com/) do
-        ssh.connect
+        ssh.with_connection(make_target(conf: no_host_key_check)) {}
       end
     end
 
     it "returns Node::ConnectError if the node name can't be resolved" do
       # even with default timeout, name resolution fails in < 1
-      ssh = Bolt::SSH.new(target(h: 'totally-not-there'))
       exec_time = Time.now
       expect_node_error(Bolt::Node::ConnectError,
                         'CONNECT_ERROR',
                         /Failed to connect to/) do
-        ssh.connect
+        ssh.with_connection(make_target(h: 'totally-not-there')) {}
       end
       exec_time = Time.now - exec_time
       expect(exec_time).to be < 1
@@ -101,12 +98,11 @@ BASH
 
     it "returns Node::ConnectError if the connection is refused" do
       # even with default timeout, connection refused fails in < 1
-      ssh = Bolt::SSH.new(target(h: hostname, p: 65535))
       exec_time = Time.now
       expect_node_error(Bolt::Node::ConnectError,
                         'CONNECT_ERROR',
                         /Failed to connect to/) do
-        ssh.connect
+        ssh.with_connection(make_target(h: hostname, p: 65535)) {}
       end
       exec_time = Time.now - exec_time
       expect(exec_time).to be < 1
@@ -119,7 +115,7 @@ BASH
       expect_node_error(Bolt::Node::ConnectError,
                         'CONNECT_ERROR',
                         /Timeout after \d+ seconds connecting to/) do
-        ssh.connect
+        ssh.with_connection(target) {}
       end
     end
 
@@ -128,11 +124,10 @@ BASH
         port = server.addr[1]
 
         timeout = mk_config(connect_timeout: 2, user: 'bad', password: 'password')
-        ssh = Bolt::SSH.new(target(h: hostname, p: port, conf: timeout))
 
         exec_time = Time.now
         expect {
-          ssh.connect
+          ssh.with_connection(make_target(h: hostname, p: port, conf: timeout)) {}
         }.to raise_error(Bolt::Node::ConnectError)
         expect(Time.now - exec_time).to be > 2
       end
@@ -142,56 +137,52 @@ BASH
   context "when executing with private key" do
     let(:config) { mk_config(host_key_check: false, key: key, user: user, port: port) }
 
-    before(:each) { ssh.connect }
-    after(:each) { ssh.disconnect }
-
     it "executes a command on a host", ssh: true do
-      expect(ssh.execute(command).stdout.string).to eq("/home/#{user}\n")
+      expect(ssh.run_command(target, command).value['stdout']).to eq("/home/#{user}\n")
     end
 
     it "can upload a file to a host", ssh: true do
+      target = make_target(conf: config)
+
       contents = "kljhdfg"
       remote_path = '/tmp/upload-test'
       with_tempfile_containing('upload-test', contents) do |file|
         expect(
-          ssh.upload(file.path, remote_path).value
+          ssh.upload(target, file.path, remote_path).value
         ).to eq(
           '_output' => "Uploaded '#{file.path}' to '#{hostname}:#{remote_path}'"
         )
 
         expect(
-          ssh.execute("cat #{remote_path}").stdout.string
+          ssh.run_command(target, "cat #{remote_path}").value['stdout']
         ).to eq(contents)
 
-        ssh.execute("rm #{remote_path}")
+        ssh.run_command(target, "rm #{remote_path}")
       end
     end
   end
 
   context "when executing" do
-    let(:ssh) { Bolt::SSH.new(target(conf: no_host_key_check)) }
-
-    before(:each) { ssh.connect }
-    after(:each) { ssh.disconnect }
+    let(:target) { make_target(conf: no_host_key_check) }
 
     it "executes a command on a host", ssh: true do
-      expect(ssh.run_command(command).value).to eq(result_value("/home/#{user}\n"))
+      expect(ssh.run_command(target, command).value).to eq(result_value("/home/#{user}\n"))
     end
 
     it "captures stderr from a host", ssh: true do
-      expect(ssh.run_command("ssh -V").value['stderr']).to match(/OpenSSH/)
+      expect(ssh.run_command(target, "ssh -V").value['stderr']).to match(/OpenSSH/)
     end
 
     it "can upload a file to a host", ssh: true do
       contents = "kljhdfg"
       with_tempfile_containing('upload-test', contents) do |file|
-        ssh.upload(file.path, "/home/#{user}/upload-test")
+        ssh.upload(target, file.path, "/home/#{user}/upload-test")
 
         expect(
-          ssh.run_command("cat /home/#{user}/upload-test")['stdout']
+          ssh.run_command(target, "cat /home/#{user}/upload-test")['stdout']
         ).to eq(contents)
 
-        ssh.execute("rm /home/#{user}/upload-test")
+        ssh.run_command(target, "rm /home/#{user}/upload-test")
       end
     end
 
@@ -199,7 +190,7 @@ BASH
       contents = "#!/bin/sh\necho hellote"
       with_tempfile_containing('script test', contents) do |file|
         expect(
-          ssh.run_script(file.path, [])['stdout']
+          ssh.run_script(target, file.path, [])['stdout']
         ).to eq("hellote\n")
       end
     end
@@ -207,19 +198,18 @@ BASH
     it "can run a script remotely with quoted arguments", ssh: true do
       with_tempfile_containing('script-test-ssh-quotes', echo_script) do |file|
         expect(
-          ssh.run_script(
-            file.path,
-            ['nospaces',
-             'with spaces',
-             "\"double double\"",
-             "'double single'",
-             '\'single single\'',
-             '"single double"',
-             "double \"double\" double",
-             "double 'single' double",
-             'single "double" single',
-             'single \'single\' single']
-          )['stdout']
+          ssh.run_script(target,
+                         file.path,
+                         ['nospaces',
+                          'with spaces',
+                          "\"double double\"",
+                          "'double single'",
+                          '\'single single\'',
+                          '"single double"',
+                          "double \"double\" double",
+                          "double 'single' double",
+                          'single "double" single',
+                          'single \'single\' single'])['stdout']
         ).to eq(<<QUOTED)
 nospaces
 with spaces
@@ -238,10 +228,9 @@ QUOTED
     it "escapes unsafe shellwords in arguments", ssh: true do
       with_tempfile_containing('script-test-ssh-escape', echo_script) do |file|
         expect(
-          ssh.run_script(
-            file.path,
-            ['echo $HOME; cat /etc/passwd']
-          )['stdout']
+          ssh.run_script(target,
+                         file.path,
+                         ['echo $HOME; cat /etc/passwd'])['stdout']
         ).to eq(<<SHELLWORDS)
 echo $HOME; cat /etc/passwd
 SHELLWORDS
@@ -252,7 +241,7 @@ SHELLWORDS
       contents = "#!/bin/sh\necho -n ${PT_message_one} ${PT_message_two}"
       arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
       with_tempfile_containing('tasks test', contents) do |file|
-        expect(ssh.run_task(file.path, 'environment', arguments).message)
+        expect(ssh.run_task(target, file.path, 'environment', arguments).message)
           .to eq('Hello from task Goodbye')
       end
     end
@@ -260,9 +249,9 @@ SHELLWORDS
     it "doesn't generate a task wrapper when not needed", ssh: true do
       contents = "#!/bin/sh\necho -n ${PT_message_one} ${PT_message_two}"
       arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
+      expect_any_instance_of(Bolt::Transport::SSH::Connection).not_to receive(:make_wrapper_stringio)
       with_tempfile_containing('tasks test', contents) do |file|
-        expect(ssh).not_to receive(:make_wrapper_stringio)
-        ssh.run_task(file.path, 'environment', arguments)
+        ssh.run_task(target, file.path, 'environment', arguments)
       end
     end
 
@@ -270,7 +259,7 @@ SHELLWORDS
       contents = "#!/bin/sh\ngrep 'message_one'"
       arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
       with_tempfile_containing('tasks test stdin', contents) do |file|
-        expect(ssh.run_task(file.path, 'stdin', arguments).value)
+        expect(ssh.run_task(target, file.path, 'stdin', arguments).value)
           .to eq("message_one" => "Hello from task", "message_two" => "Goodbye")
       end
     end
@@ -283,7 +272,7 @@ grep 'message_one'
 SHELL
       arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
       with_tempfile_containing('tasks-test-both', contents) do |file|
-        expect(ssh.run_task(file.path, 'both', arguments).message).to eq(<<SHELL)
+        expect(ssh.run_task(target, file.path, 'both', arguments).message).to eq(<<SHELL)
 Hello from task Goodbye{\"message_one\":\
 \"Hello from task\",\"message_two\":\"Goodbye\"}
 SHELL
@@ -292,7 +281,7 @@ SHELL
 
     context "when it can't upload a file" do
       before(:each) do
-        expect(ssh).to receive(:write_remote_file).and_raise(
+        allow_any_instance_of(Bolt::Transport::SSH::Connection).to receive(:write_remote_file).and_raise(
           Bolt::Node::FileError.new("no write", "WRITE_ERROR")
         )
       end
@@ -301,7 +290,7 @@ SHELL
         contents = "kljhdfg"
         with_tempfile_containing('upload-test', contents) do |file|
           expect {
-            ssh.upload(file.path, "/home/#{user}/upload-test")
+            ssh.upload(target, file.path, "/home/#{user}/upload-test")
           }.to raise_error(Bolt::Node::FileError, 'no write')
         end
       end
@@ -310,7 +299,7 @@ SHELL
         contents = "#!/bin/sh\necho hellote"
         with_tempfile_containing('script test', contents) do |file|
           expect {
-            ssh.run_script(file.path, [])
+            ssh.run_script(target, file.path, [])
           }.to raise_error(Bolt::Node::FileError, 'no write')
         end
       end
@@ -320,7 +309,7 @@ SHELL
         arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
         with_tempfile_containing('tasks test', contents) do |file|
           expect {
-            ssh.run_task(file.path, 'environment', arguments)
+            ssh.run_task(target, file.path, 'environment', arguments)
           }.to raise_error(Bolt::Node::FileError, 'no write')
         end
       end
@@ -328,7 +317,7 @@ SHELL
 
     context "when it can't create a tempfile" do
       before(:each) do
-        expect(ssh).to receive(:make_tempdir).and_raise(
+        allow_any_instance_of(Bolt::Transport::SSH::Connection).to receive(:make_tempdir).and_raise(
           Bolt::Node::FileError.new("no tmpdir", "TEMDIR_ERROR")
         )
       end
@@ -337,7 +326,7 @@ SHELL
         contents = "#!/bin/sh\necho hellote"
         with_tempfile_containing('script test', contents) do |file|
           expect {
-            ssh.run_script(file.path, []).error_hash['msg']
+            ssh.run_script(target, file.path, []).error_hash['msg']
           }.to raise_error(Bolt::Node::FileError, 'no tmpdir')
         end
       end
@@ -347,7 +336,7 @@ SHELL
         arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
         with_tempfile_containing('tasks test', contents) do |file|
           expect {
-            ssh.run_task(file.path, 'environment', arguments)
+            ssh.run_task(target, file.path, 'environment', arguments)
           }.to raise_error(Bolt::Node::FileError, 'no tmpdir')
         end
       end
@@ -358,29 +347,24 @@ SHELL
     let(:tmpdir) { '/tmp/mytempdir' }
     let(:config) { mk_config(host_key_check: false, tmpdir: tmpdir, user: user, password: password) }
 
-    before(:each) { ssh.connect }
     after(:each) do
-      begin
-        ssh.run_command("rm -rf #{tmpdir}")
-      ensure
-        ssh.disconnect
-      end
+      ssh.run_command(target, "rm -rf #{tmpdir}")
     end
 
     it "errors when tmpdir doesn't exist", ssh: true do
       contents = "#!/bin/sh\n echo $0"
       with_tempfile_containing('script dir', contents) do |file|
         expect {
-          ssh.run_script(file.path, [])
+          ssh.run_script(target, file.path, [])
         }.to raise_error(Bolt::Node::FileError, /Could not make tempdir.*#{Regexp.escape(tmpdir)}/)
       end
     end
 
     it 'uploads a script to the specified tmpdir', ssh: true do
-      ssh.run_command("mkdir #{tmpdir}")
+      ssh.run_command(target, "mkdir #{tmpdir}")
       contents = "#!/bin/sh\n echo $0"
       with_tempfile_containing('script dir', contents) do |file|
-        expect(ssh.run_script(file.path, [])['stdout']).to match(/#{Regexp.escape(tmpdir)}/)
+        expect(ssh.run_script(target, file.path, [])['stdout']).to match(/#{Regexp.escape(tmpdir)}/)
       end
     end
   end
@@ -390,18 +374,15 @@ SHELL
       mk_config(host_key_check: false, sudo_password: password, run_as: 'root', user: user, password: password)
     }
 
-    before(:each) { ssh.connect }
-    after(:each) { ssh.disconnect }
-
     it "can execute a command", ssh: true do
-      expect(ssh.run_command('whoami')['stdout']).to eq("root\n")
+      expect(ssh.run_command(target, 'whoami')['stdout']).to eq("root\n")
     end
 
     it "can run a task passing input on stdin", ssh: true do
       contents = "#!/bin/sh\ngrep 'message_one'"
       arguments = { message_one: 'Hello from task', message_two: 'Goodbye' }
       with_tempfile_containing('tasks test stdin', contents) do |file|
-        expect(ssh.run_task(file.path, 'stdin', arguments).value)
+        expect(ssh.run_task(target, file.path, 'stdin', arguments).value)
           .to eq("message_one" => "Hello from task", "message_two" => "Goodbye")
       end
     end
@@ -410,12 +391,12 @@ SHELL
       contents = "upload file test as root content"
       dest = '/tmp/root-file-upload-test'
       with_tempfile_containing('tasks test upload as root', contents) do |file|
-        expect(ssh.upload(file.path, dest).message).to match(/Uploaded/)
-        expect(ssh.run_command("cat #{dest}")['stdout']).to eq(contents)
-        expect(ssh.run_command("stat -c %U #{dest}")['stdout'].chomp).to eq('root')
+        expect(ssh.upload(target, file.path, dest).message).to match(/Uploaded/)
+        expect(ssh.run_command(target, "cat #{dest}")['stdout']).to eq(contents)
+        expect(ssh.run_command(target, "stat -c %U #{dest}")['stdout'].chomp).to eq('root')
       end
 
-      ssh.execute("rm #{dest}", sudoable: true, run_as: 'root')
+      ssh.run_command(target, "rm #{dest}", sudoable: true, run_as: 'root')
     end
 
     context "requesting a pty" do
@@ -425,7 +406,7 @@ SHELL
       }
 
       it "can execute a command when a tty is requested", ssh: true do
-        expect(ssh.run_command('whoami')['stdout']).to eq("\r\nroot\r\n")
+        expect(ssh.run_command(target, 'whoami')['stdout']).to eq("\r\nroot\r\n")
       end
     end
 
@@ -435,20 +416,20 @@ SHELL
       }
 
       it "can override run_as for command via an option", ssh: true do
-        expect(ssh.run_command('whoami', '_run_as' => 'root')['stdout']).to eq("root\n")
+        expect(ssh.run_command(target, 'whoami', '_run_as' => 'root')['stdout']).to eq("root\n")
       end
 
       it "can override run_as for script via an option", ssh: true do
         contents = "#!/bin/sh\nwhoami"
         with_tempfile_containing('script test', contents) do |file|
-          expect(ssh.run_script(file.path, [], '_run_as' => 'root')['stdout']).to eq("root\n")
+          expect(ssh.run_script(target, file.path, [], '_run_as' => 'root')['stdout']).to eq("root\n")
         end
       end
 
       it "can override run_as for task via an option", ssh: true do
         contents = "#!/bin/sh\nwhoami"
         with_tempfile_containing('tasks test', contents) do |file|
-          expect(ssh.run_task(file.path, 'environment', {}, '_run_as' => 'root').message).to eq("root\n")
+          expect(ssh.run_task(target, file.path, 'environment', {}, '_run_as' => 'root').message).to eq("root\n")
         end
       end
 
@@ -456,12 +437,12 @@ SHELL
         contents = "upload file test as root content"
         dest = '/tmp/root-file-upload-test'
         with_tempfile_containing('tasks test upload as root', contents) do |file|
-          expect(ssh.upload(file.path, dest, '_run_as' => 'root').message).to match(/Uploaded/)
-          expect(ssh.run_command("cat #{dest}", '_run_as' => 'root')['stdout']).to eq(contents)
-          expect(ssh.run_command("stat -c %U #{dest}", '_run_as' => 'root')['stdout'].chomp).to eq('root')
+          expect(ssh.upload(target, file.path, dest, '_run_as' => 'root').message).to match(/Uploaded/)
+          expect(ssh.run_command(target, "cat #{dest}", '_run_as' => 'root')['stdout']).to eq(contents)
+          expect(ssh.run_command(target, "stat -c %U #{dest}", '_run_as' => 'root')['stdout'].chomp).to eq('root')
         end
 
-        ssh.execute("rm #{dest}", sudoable: true, run_as: 'root')
+        ssh.run_command(target, "rm #{dest}", sudoable: true, run_as: 'root')
       end
     end
 
@@ -473,7 +454,7 @@ SHELL
 
       it "returns a failed result", ssh: true do
         expect {
-          ssh.run_command('whoami')
+          ssh.run_command(target, 'whoami')
              .to raise_error(Bolt::Node::EscalateError,
                              "Sudo password for user #{user} not recognized on #{hostname}:#{port}")
         }
@@ -485,7 +466,7 @@ SHELL
 
       it "returns a failed result", ssh: true do
         expect {
-          ssh.run_command('whoami')
+          ssh.run_command(target, 'whoami')
              .to raise_error(Bolt::Node::EscalateError,
                              "Sudo password for user #{user} was not provided for #{hostname}:#{port}")
         }
